pyproject.toml
[project]
name = "structured"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "zzstoatzz", email = "thrast36@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic-ai-slim[openai]",
    "pydantic-settings",
]


[project.optional-dependencies]
dev = ["ruff", "ipython"]

[tool.ruff]
line-length = 120
target-version = "py39"


[tool.ruff.lint]
extend-select = ["Q", "RUF100", "C90", "UP", "I", "D"]
flake8-quotes = { inline-quotes = "single", multiline-quotes = "double" }
isort = { combine-as-imports = true }
mccabe = { max-complexity = 15 }
ignore = ["D415"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"


[project.scripts]
structured = "structured:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
